mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain-mcp-adapters>=0.1.9",
    "mcp>=1.11.0",
    "openai>=1.93.3",
    "python-dotenv>=1.1.1",
]
